requirements_pytorch.txt
# Pytorch requirements
torch
torchvision
numpy
pillow
tqdm
scikit-learn
rasterio
matplotlib